requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
Django==2.2.5
django-cors-headers==3.11.0
django-countries==7.3.2
django-filter==21.1
django-seed==0.3.1
djangorestframework==3.13.1
Faker==14.1.0
flake8==5.0.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
toposort==1.7
typing_extensions==4.3.0
asgiref==3.5.2
black==22.6.0
click==8.1.3
Django==2.2.5
django-cors-headers==3.11.0
django-countries==7.3.2
django-filter==21.1
django-seed==0.3.1
djangorestframework==3.13.1
Faker==14.1.0
flake8==5.0.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
toposort==1.7
typing_extensions==4.3.0
